+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 - Add SingleLineThrowFixer
 - Add PhpUnitNoUselessReturnFixer
+- Feature: NoCommentedOutCodeFixer - handle class method
 
 ## v1.13.0 - *2019-01-11*
 - Update PHP CS Fixer to v2.14